fix(data-integrity): Sanitize.integer destruía punto decimal → inflación 100x en sqft
Root cause: gsub(/[^\d]/) eliminaba el '.' de valores formateados
como '1,200.00' produciendo '120000' (inflación 100x).

Cambios:
- sanitize.rb: integer() usa /[^\d.\-]/ + to_d.round(0) para preservar decimales
- sheet_row_processor.rb: sqft_lot/sqft_living → Sanitize.decimal (alineado con decimal() en PostgreSQL)
- lib/tasks/fix_inflated_sqft.rb: script de auditoría retroactiva

Validación post-sync:
- sqft_living corruptos: 152 → 0 (100% reparado)
- sqft_lot corruptos: 1367 → 125 (legítimos, terrenos rurales)
- Distribución normalizada: curva residencial natural 500-3000 sqft

CRM Immunity: parcel_user_tags y parcel_user_notes intactos.

+++ app/services/sheet_row_processor.rb
@@ -106,10 +106,10 @@ class SheetRowProcessor
       estimated_sale_value: Sanitize.currency(col(ESTIMATED_SALE_VALUE)),
       price_per_acre:       calculate_price_per_acre,
 
-      # ── NUMÉRICOS ────────────────────────────────────────────────────
+      # ── NUMÉRICOS (Sanitize.decimal — alineado con decimal() en PostgreSQL) ──
       lot_area_acres:       Sanitize.decimal(col(LOT_AREA_ACRES)),
-      sqft_lot:             Sanitize.integer(col(SQFT_LOT)),
-      sqft_living:          Sanitize.integer(col(SQFT_LIVING)),
+      sqft_lot:             Sanitize.decimal(col(SQFT_LOT)),
+      sqft_living:          Sanitize.decimal(col(SQFT_LIVING)),
 
       # ── TEXTO (Sanitize.text) ─────────────────────────────────────────
       address:              Sanitize.text(col(ADDRESS)),
